sorting moved to a single component
