+++ src/views/TasksView.vue
@@ -6,8 +6,7 @@ import CreateTask from '@/components/CreateTask.vue'
 import TaskCard from '@/components/TaskCard.vue'
 import { useUserStore } from '@/stores/userStore'
 import { useRouter } from 'vue-router'
-import SortComponent from '@/components/SortComponent.vue'
-import SortingModal from '@/components/SortingModal.vue'
+import SortTasks from '@/components/SortTasks.vue'
 
 const router = useRouter()
 const tasksStore = useTasksStore()
@@ -34,14 +33,7 @@ onMounted(() => {
     <template v-if="tasks && tasks.length">
       <p class="mb-2">Total tasks: {{ tasks.length }}</p>
       <div class="sorting-tools flex justify-between">
-        <!-- SortComponent for desktop -->
-        <div class="hidden sm:block">
-          <SortComponent />
-        </div>
-        <!-- SortingModal for mobile -->
-        <div class="sm:hidden">
-          <SortingModal />
-        </div>
+        <SortTasks />
       </div>
       <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
         <TaskCard v-for="task in tasks" :key="task.id" :task="task" />
